CleanUp and readme 2.0
Clean Up codes AND add more usage details in readme
